Remove V4 naming prefix from all classes, files, CLI functions, and tests (#162)
Comprehensive V4 naming cleanup across the entire codebase:

- Rename 34+ classes: V4Workspace→Workspace, V4Config→Config, V4Strategy→Strategy,
  V4Verifier→Verifier, V4Validator→Validator, V4Runner→Runner, V4IngestProcessor→
  IngestProcessor, V4CodeGenerator→CodeGenerator, etc.
- Rename files: v4_processor.py→strategy_processor.py, v4_generator.py→
  strategy_generator.py, v4_runner.py→runner.py (with forwarding modules at old paths)
- Rename 13 CLI functions: cmd_v4_ingest→cmd_ingest, cmd_v4_verify→cmd_verify, etc.
- Add parent-level re-exports in core/__init__.py and schemas/__init__.py
- Update all 15 test files to use new names (490 tests passing)
- Fix 3 duplicate function definition bugs in main.py (cmd_status, cmd_run, cmd_ideate)
- All backward-compat aliases preserved (V4Name = Name pattern)
