+++ research_system/cli/main.py
@@ -118,7 +118,7 @@ For more information, visit: https://github.com/your-repo/research-system
     _add_migrate_parser(subparsers)
 
     # Add main commands (formerly V4)
-    _add_v4_commands(subparsers)
+    _add_commands(subparsers)
 
     # Add update command
     _add_update_parser(subparsers)
@@ -772,7 +772,7 @@ def _add_migrate_parser(subparsers):
 # ============================================================================
 
 
-def _add_v4_commands(subparsers):
+def _add_commands(subparsers):
     """Add V4-related subcommands.
 
     V4 commands support the new research workflow:
@@ -822,7 +822,7 @@ Strategies meeting minimum thresholds are created in strategies/pending/.
         action="store_true",
         help="Skip quality checks and create strategies anyway (useful for testing without API key)"
     )
-    parser.set_defaults(func=cmd_v4_ingest)
+    parser.set_defaults(func=cmd_ingest)
 
     # verify command
     parser = subparsers.add_parser(
@@ -862,7 +862,7 @@ Tests include:
         dest="process_all",
         help="Verify all pending strategies"
     )
-    parser.set_defaults(func=cmd_v4_verify)
+    parser.set_defaults(func=cmd_verify)
 
     # validate command
     parser = subparsers.add_parser(
@@ -909,7 +909,7 @@ strategy robustness across different market regimes.
         dest="process_all",
         help="Validate all strategies that have verification results"
     )
-    parser.set_defaults(func=cmd_v4_validate)
+    parser.set_defaults(func=cmd_validate)
 
     # learn command
     parser = subparsers.add_parser(
@@ -948,7 +948,7 @@ Analyzes validation results to identify:
         dest="process_all",
         help="Extract learnings from all strategies with validation results"
     )
-    parser.set_defaults(func=cmd_v4_learn)
+    parser.set_defaults(func=cmd_learn)
 
     # ideate command
     parser = subparsers.add_parser(
@@ -981,7 +981,7 @@ Ideas are generated by:
         action="store_true",
         help="Show ideas without saving"
     )
-    parser.set_defaults(func=cmd_v4_ideate)
+    parser.set_defaults(func=cmd_ideate)
 
     # run command
     parser = subparsers.add_parser(
@@ -1052,7 +1052,7 @@ Examples:
         metavar="PATH",
         help="Path to workspace"
     )
-    parser.set_defaults(func=cmd_v4_run)
+    parser.set_defaults(func=cmd_run)
 
     # cleanup command
     parser = subparsers.add_parser(
@@ -1087,7 +1087,7 @@ Examples:
         metavar="PATH",
         help="Path to workspace"
     )
-    parser.set_defaults(func=cmd_v4_cleanup)
+    parser.set_defaults(func=cmd_cleanup)
 
     # walkforward command
     parser = subparsers.add_parser(
@@ -1162,7 +1162,7 @@ Examples:
         metavar="PATH",
         help="Path to workspace"
     )
-    parser.set_defaults(func=cmd_v4_walkforward)
+    parser.set_defaults(func=cmd_walkforward)
 
     # status command
     parser = subparsers.add_parser(
@@ -1185,7 +1185,7 @@ Displays a summary of the workspace including:
         metavar="PATH",
         help="Path to workspace"
     )
-    parser.set_defaults(func=cmd_v4_status)
+    parser.set_defaults(func=cmd_status)
 
     # list command
     parser = subparsers.add_parser(
@@ -1224,7 +1224,7 @@ Statuses:
         metavar="PATH",
         help="Path to workspace"
     )
-    parser.set_defaults(func=cmd_v4_list)
+    parser.set_defaults(func=cmd_list)
 
     # show command
     parser = subparsers.add_parser(
@@ -1258,7 +1258,7 @@ Shows all strategy metadata including:
         metavar="PATH",
         help="Path to workspace"
     )
-    parser.set_defaults(func=cmd_v4_show)
+    parser.set_defaults(func=cmd_show)
 
     # config command
     parser = subparsers.add_parser(
@@ -1286,7 +1286,7 @@ Configuration includes:
         metavar="PATH",
         help="Path to workspace"
     )
-    parser.set_defaults(func=cmd_v4_config)
+    parser.set_defaults(func=cmd_config)
 
 
 # ============================================================================
@@ -1349,7 +1349,7 @@ def cmd_init_v4(args):
 # ============================================================================
 
 
-def cmd_v4_ingest(args):
+def cmd_ingest(args):
     """Ingest files from inbox into strategies ."""
     workspace = get_workspace_from_args(args)
 
@@ -1492,7 +1492,7 @@ def cmd_v4_ingest(args):
     return 0
 
 
-def _cmd_v4_verify_all(workspace, args):
+def _cmd_verify_all(workspace, args):
     """Verify all pending strategies."""
     from research_system.validation import V4Verifier, VerificationStatus
 
@@ -1551,7 +1551,7 @@ def _cmd_v4_verify_all(workspace, args):
     return 0
 
 
-def _cmd_v4_validate_all(workspace, args):
+def _cmd_validate_all(workspace, args):
     """Validate all strategies that have verification results."""
     from research_system.validation import V4Verifier, V4Validator, VerificationStatus, GateStatus
 
@@ -1644,7 +1644,7 @@ def _cmd_v4_validate_all(workspace, args):
     return 0
 
 
-def _cmd_v4_learn_all(workspace, args):
+def _cmd_learn_all(workspace, args):
     """Extract learnings from all strategies with validation results."""
     from research_system.validation import V4Learner
 
@@ -1714,7 +1714,7 @@ def _cmd_v4_learn_all(workspace, args):
     return 0
 
 
-def cmd_v4_verify(args):
+def cmd_verify(args):
     """Run verification tests on a strategy ."""
     from research_system.validation import V4Verifier, VerificationStatus
 
@@ -1732,7 +1732,7 @@ def cmd_v4_verify(args):
 
     # Handle --all flag
     if process_all:
-        return _cmd_v4_verify_all(workspace, args)
+        return _cmd_verify_all(workspace, args)
 
     if not strategy_id:
         print("Error: Strategy ID required")
@@ -1792,7 +1792,7 @@ def cmd_v4_verify(args):
     return 0 if result.overall_status != VerificationStatus.FAIL else 1
 
 
-def cmd_v4_validate(args):
+def cmd_validate(args):
     """Run walk-forward validation on a strategy ."""
     import json
     import yaml
@@ -1817,7 +1817,7 @@ def cmd_v4_validate(args):
 
     # Handle --all flag
     if process_all:
-        return _cmd_v4_validate_all(workspace, args)
+        return _cmd_validate_all(workspace, args)
 
     if not strategy_id:
         print("Error: Strategy ID required")
@@ -1945,7 +1945,7 @@ def cmd_v4_validate(args):
     return 0 if result.overall_passed or not backtest_results else 1
 
 
-def cmd_v4_learn(args):
+def cmd_learn(args):
     """Extract learnings from validation results ."""
     from research_system.validation import V4Learner
 
@@ -1963,7 +1963,7 @@ def cmd_v4_learn(args):
 
     # Handle --all flag
     if process_all:
-        return _cmd_v4_learn_all(workspace, args)
+        return _cmd_learn_all(workspace, args)
 
     if not strategy_id:
         print("Error: Strategy ID required")
@@ -2042,7 +2042,7 @@ def cmd_v4_learn(args):
     return 0
 
 
-def cmd_v4_ideate(args):
+def cmd_ideate(args):
     """Generate new strategy ideas ."""
     from research_system.validation import V4Ideator, V4Learner
 
@@ -2116,7 +2116,7 @@ def cmd_v4_ideate(args):
     return 0
 
 
-def cmd_v4_run(args):
+def cmd_run(args):
     """Run full validation pipeline ."""
     from research_system.validation.v4_runner import V4Runner
 
@@ -2194,7 +2194,7 @@ def cmd_v4_run(args):
         return 0
 
 
-def cmd_v4_cleanup(args):
+def cmd_cleanup(args):
     """Clean up stuck QC backtests ."""
     from research_system.validation.backtest import BacktestExecutor
 
@@ -2250,7 +2250,7 @@ def cmd_v4_cleanup(args):
     return 0
 
 
-def cmd_v4_walkforward(args):
+def cmd_walkforward(args):
     """Run true walk-forward optimization ."""
     from research_system.optimization import (
         WalkForwardConfig,
@@ -2350,7 +2350,7 @@ def cmd_v4_walkforward(args):
     return 0 if result.success else 1
 
 
-def cmd_v4_status(args):
+def cmd_status(args):
     """Show workspace status dashboard ."""
     workspace = get_workspace_from_args(args)
 
@@ -2445,7 +2445,7 @@ def cmd_v4_status(args):
     return 0
 
 
-def cmd_v4_list(args):
+def cmd_list(args):
     """List strategies ."""
     workspace = get_workspace_from_args(args)
 
@@ -2512,7 +2512,7 @@ def cmd_v4_list(args):
     return 0
 
 
-def cmd_v4_show(args):
+def cmd_show(args):
     """Show strategy details ."""
     import yaml
 
@@ -2747,7 +2747,7 @@ def _print_strategy_details(strategy: dict) -> None:
     print()
 
 
-def cmd_v4_config(args):
+def cmd_config(args):
     """Show/validate V4 configuration."""
     workspace = get_workspace_from_args(args)
 
@@ -3561,7 +3561,7 @@ def cmd_validate_check(args):
         return 0
 
 
-def cmd_status(args):
+def cmd_status_reports(args):
     """View dashboard and reports."""
     import subprocess
     from scripts.status.generate_reports import (
@@ -4185,7 +4185,7 @@ def cmd_analyze_show(args):
     return 0
 
 
-def cmd_ideate(args):
+def cmd_ideate_catalog(args):
     """Generate strategy ideas using multiple personas."""
     ws = require_workspace(args.workspace)
 
@@ -4484,7 +4484,7 @@ def cmd_migrate_master_index(args):
 # Run Command - The Core Validation + Expert Loop
 # ============================================================================
 
-def cmd_run(args):
+def cmd_run_catalog(args):
     """Run the full validation + expert review loop."""
     ws = require_workspace(args.workspace)
     catalog = Catalog(ws.catalog_path)
