+++ research_system/validation/__init__.py
@@ -1,25 +1,25 @@
 """V4 Validation module for strategy verification and walk-forward testing."""
 
 from research_system.validation.verifier import (
-    V4Verifier,
+    Verifier,
     VerificationResult,
     VerificationTest,
     VerificationStatus,
 )
 from research_system.validation.validator import (
-    V4Validator,
+    Validator,
     ValidationResult,
     ValidationGate,
     GateStatus,
     GateResult,
 )
 from research_system.validation.learner import (
-    V4Learner,
+    Learner,
     LearningsDocument,
     Learning,
 )
 from research_system.validation.ideator import (
-    V4Ideator,
+    Ideator,
     StrategyIdea,
 )
 from research_system.validation.backtest import (
@@ -28,28 +28,40 @@ from research_system.validation.backtest import (
     WalkForwardResult,
     WalkForwardWindow,
 )
-from research_system.validation.v4_runner import (
-    V4Runner,
-    V4RunResult,
+from research_system.validation.runner import (
+    Runner,
+    RunResult,
 )
+
+# Backward-compat aliases
+V4Verifier = Verifier
+V4Validator = Validator
+V4Learner = Learner
+V4Ideator = Ideator
+V4Runner = Runner
+V4RunResult = RunResult
 
 __all__ = [
     # Verifier
+    "Verifier",
     "V4Verifier",
     "VerificationResult",
     "VerificationTest",
     "VerificationStatus",
     # Validator
+    "Validator",
     "V4Validator",
     "ValidationResult",
     "ValidationGate",
     "GateStatus",
     "GateResult",
     # Learner
+    "Learner",
     "V4Learner",
     "LearningsDocument",
     "Learning",
     # Ideator
+    "Ideator",
     "V4Ideator",
     "StrategyIdea",
     # Backtest
@@ -58,6 +70,8 @@ __all__ = [
     "WalkForwardResult",
     "WalkForwardWindow",
     # Runner
+    "Runner",
     "V4Runner",
+    "RunResult",
     "V4RunResult",
 ]
