+++ tests/v4/test_cli.py
@@ -1,10 +1,10 @@
-"""Tests for V4 CLI commands.
+"""Tests for CLI commands.
 
-This module tests the V4 CLI commands:
-1. Test --help output includes all V4 commands
+This module tests the CLI commands:
+1. Test --help output includes all commands
 2. Test each command prints "not implemented" message
 3. Test --workspace flag is accepted
-4. Test research init --v4 creates workspace structure
+4. Test research init creates workspace structure
 """
 
 import pytest
@@ -18,11 +18,11 @@ from pathlib import Path
 # =============================================================================
 
 
-class TestV4CLIHelp:
-    """Test that V4 commands appear in help output."""
+class TestCLIHelp:
+    """Test that commands appear in help output."""
 
-    def test_main_help_shows_v4_ingest(self):
-        """Test main help shows v4-ingest command."""
+    def test_main_help_shows_ingest(self):
+        """Test main help shows ingest command."""
         result = subprocess.run(
             [sys.executable, "-m", "research_system.cli.main", "--help"],
             capture_output=True,
@@ -30,10 +30,10 @@ class TestV4CLIHelp:
             cwd=Path(__file__).parent.parent.parent
         )
         assert result.returncode == 0
-        assert "v4-ingest" in result.stdout
+        assert "ingest" in result.stdout
 
-    def test_main_help_shows_v4_verify(self):
-        """Test main help shows v4-verify command."""
+    def test_main_help_shows_verify(self):
+        """Test main help shows verify command."""
         result = subprocess.run(
             [sys.executable, "-m", "research_system.cli.main", "--help"],
             capture_output=True,
@@ -41,10 +41,10 @@ class TestV4CLIHelp:
             cwd=Path(__file__).parent.parent.parent
         )
         assert result.returncode == 0
-        assert "v4-verify" in result.stdout
+        assert "verify" in result.stdout
 
-    def test_main_help_shows_v4_validate(self):
-        """Test main help shows v4-validate command."""
+    def test_main_help_shows_validate(self):
+        """Test main help shows validate command."""
         result = subprocess.run(
             [sys.executable, "-m", "research_system.cli.main", "--help"],
             capture_output=True,
@@ -52,10 +52,10 @@ class TestV4CLIHelp:
             cwd=Path(__file__).parent.parent.parent
         )
         assert result.returncode == 0
-        assert "v4-validate" in result.stdout
+        assert "validate" in result.stdout
 
-    def test_main_help_shows_v4_learn(self):
-        """Test main help shows v4-learn command."""
+    def test_main_help_shows_learn(self):
+        """Test main help shows learn command."""
         result = subprocess.run(
             [sys.executable, "-m", "research_system.cli.main", "--help"],
             capture_output=True,
@@ -63,10 +63,10 @@ class TestV4CLIHelp:
             cwd=Path(__file__).parent.parent.parent
         )
         assert result.returncode == 0
-        assert "v4-learn" in result.stdout
+        assert "learn" in result.stdout
 
-    def test_main_help_shows_v4_status(self):
-        """Test main help shows v4-status command."""
+    def test_main_help_shows_status(self):
+        """Test main help shows status command."""
         result = subprocess.run(
             [sys.executable, "-m", "research_system.cli.main", "--help"],
             capture_output=True,
@@ -74,10 +74,10 @@ class TestV4CLIHelp:
             cwd=Path(__file__).parent.parent.parent
         )
         assert result.returncode == 0
-        assert "v4-status" in result.stdout
+        assert "status" in result.stdout
 
-    def test_main_help_shows_v4_list(self):
-        """Test main help shows v4-list command."""
+    def test_main_help_shows_list(self):
+        """Test main help shows list command."""
         result = subprocess.run(
             [sys.executable, "-m", "research_system.cli.main", "--help"],
             capture_output=True,
@@ -85,10 +85,10 @@ class TestV4CLIHelp:
             cwd=Path(__file__).parent.parent.parent
         )
         assert result.returncode == 0
-        assert "v4-list" in result.stdout
+        assert "list" in result.stdout
 
-    def test_main_help_shows_v4_show(self):
-        """Test main help shows v4-show command."""
+    def test_main_help_shows_show(self):
+        """Test main help shows show command."""
         result = subprocess.run(
             [sys.executable, "-m", "research_system.cli.main", "--help"],
             capture_output=True,
@@ -96,10 +96,10 @@ class TestV4CLIHelp:
             cwd=Path(__file__).parent.parent.parent
         )
         assert result.returncode == 0
-        assert "v4-show" in result.stdout
+        assert "show" in result.stdout
 
-    def test_main_help_shows_v4_config(self):
-        """Test main help shows v4-config command."""
+    def test_main_help_shows_config(self):
+        """Test main help shows config command."""
         result = subprocess.run(
             [sys.executable, "-m", "research_system.cli.main", "--help"],
             capture_output=True,
@@ -107,10 +107,10 @@ class TestV4CLIHelp:
             cwd=Path(__file__).parent.parent.parent
         )
         assert result.returncode == 0
-        assert "v4-config" in result.stdout
+        assert "config" in result.stdout
 
-    def test_init_help_shows_v4_flag(self):
-        """Test init --help shows --v4 flag."""
+    def test_init_help(self):
+        """Test init --help shows expected options."""
         result = subprocess.run(
             [sys.executable, "-m", "research_system.cli.main", "init", "--help"],
             capture_output=True,
@@ -118,16 +118,16 @@ class TestV4CLIHelp:
             cwd=Path(__file__).parent.parent.parent
         )
         assert result.returncode == 0
-        assert "--v4" in result.stdout
+        assert "path" in result.stdout.lower()
 
 
-class TestV4CommandHelp:
-    """Test that each V4 command shows proper help."""
+class TestCommandHelp:
+    """Test that each command shows proper help."""
 
-    def test_v4_ingest_help(self):
-        """Test v4-ingest shows help with description."""
+    def test_ingest_help(self):
+        """Test ingest shows help with description."""
         result = subprocess.run(
-            [sys.executable, "-m", "research_system.cli.main", "v4-ingest", "--help"],
+            [sys.executable, "-m", "research_system.cli.main", "ingest", "--help"],
             capture_output=True,
             text=True,
             cwd=Path(__file__).parent.parent.parent
@@ -136,10 +136,10 @@ class TestV4CommandHelp:
         assert "inbox" in result.stdout.lower()
         assert "--workspace" in result.stdout or "-w" in result.stdout
 
-    def test_v4_verify_help(self):
-        """Test v4-verify shows help with description."""
+    def test_verify_help(self):
+        """Test verify shows help with description."""
         result = subprocess.run(
-            [sys.executable, "-m", "research_system.cli.main", "v4-verify", "--help"],
+            [sys.executable, "-m", "research_system.cli.main", "verify", "--help"],
             capture_output=True,
             text=True,
             cwd=Path(__file__).parent.parent.parent
@@ -148,10 +148,10 @@ class TestV4CommandHelp:
         assert "verification" in result.stdout.lower() or "bias" in result.stdout.lower()
         assert "--workspace" in result.stdout or "-w" in result.stdout
 
-    def test_v4_validate_help(self):
-        """Test v4-validate shows help with description."""
+    def test_validate_help(self):
+        """Test validate shows help with description."""
         result = subprocess.run(
-            [sys.executable, "-m", "research_system.cli.main", "v4-validate", "--help"],
+            [sys.executable, "-m", "research_system.cli.main", "validate", "--help"],
             capture_output=True,
             text=True,
             cwd=Path(__file__).parent.parent.parent
@@ -160,10 +160,10 @@ class TestV4CommandHelp:
         assert "walk-forward" in result.stdout.lower() or "validation" in result.stdout.lower()
         assert "--workspace" in result.stdout or "-w" in result.stdout
 
-    def test_v4_learn_help(self):
-        """Test v4-learn shows help with description."""
+    def test_learn_help(self):
+        """Test learn shows help with description."""
         result = subprocess.run(
-            [sys.executable, "-m", "research_system.cli.main", "v4-learn", "--help"],
+            [sys.executable, "-m", "research_system.cli.main", "learn", "--help"],
             capture_output=True,
             text=True,
             cwd=Path(__file__).parent.parent.parent
@@ -172,10 +172,10 @@ class TestV4CommandHelp:
         assert "learning" in result.stdout.lower()
         assert "--workspace" in result.stdout or "-w" in result.stdout
 
-    def test_v4_status_help(self):
-        """Test v4-status shows help with description."""
+    def test_status_help(self):
+        """Test status shows help with description."""
         result = subprocess.run(
-            [sys.executable, "-m", "research_system.cli.main", "v4-status", "--help"],
+            [sys.executable, "-m", "research_system.cli.main", "status", "--help"],
             capture_output=True,
             text=True,
             cwd=Path(__file__).parent.parent.parent
@@ -184,10 +184,10 @@ class TestV4CommandHelp:
         assert "status" in result.stdout.lower()
         assert "--workspace" in result.stdout or "-w" in result.stdout
 
-    def test_v4_list_help(self):
-        """Test v4-list shows help with description."""
+    def test_list_help(self):
+        """Test list shows help with description."""
         result = subprocess.run(
-            [sys.executable, "-m", "research_system.cli.main", "v4-list", "--help"],
+            [sys.executable, "-m", "research_system.cli.main", "list", "--help"],
             capture_output=True,
             text=True,
             cwd=Path(__file__).parent.parent.parent
@@ -197,10 +197,10 @@ class TestV4CommandHelp:
         assert "--status" in result.stdout
         assert "--workspace" in result.stdout or "-w" in result.stdout
 
-    def test_v4_show_help(self):
-        """Test v4-show shows help with description."""
+    def test_show_help(self):
+        """Test show shows help with description."""
         result = subprocess.run(
-            [sys.executable, "-m", "research_system.cli.main", "v4-show", "--help"],
+            [sys.executable, "-m", "research_system.cli.main", "show", "--help"],
             capture_output=True,
             text=True,
             cwd=Path(__file__).parent.parent.parent
@@ -209,10 +209,10 @@ class TestV4CommandHelp:
         assert "strategy_id" in result.stdout.lower()
         assert "--workspace" in result.stdout or "-w" in result.stdout
 
-    def test_v4_config_help(self):
-        """Test v4-config shows help with description."""
+    def test_config_help(self):
+        """Test config shows help with description."""
         result = subprocess.run(
-            [sys.executable, "-m", "research_system.cli.main", "v4-config", "--help"],
+            [sys.executable, "-m", "research_system.cli.main", "config", "--help"],
             capture_output=True,
             text=True,
             cwd=Path(__file__).parent.parent.parent
@@ -228,21 +228,21 @@ class TestV4CommandHelp:
 # =============================================================================
 
 
-class TestV4NotImplemented:
-    """Test that V4 commands print 'not implemented' message or work correctly."""
+class TestCommandBehavior:
+    """Test that commands work correctly."""
 
-    def test_v4_ingest_empty_inbox(self, tmp_path):
-        """Test v4-ingest handles empty inbox gracefully."""
-        # First init a v4 workspace
+    def test_ingest_empty_inbox(self, tmp_path):
+        """Test ingest handles empty inbox gracefully."""
+        # First init a workspace
         subprocess.run(
-            [sys.executable, "-m", "research_system.cli.main", "init", "--v4", str(tmp_path)],
+            [sys.executable, "-m", "research_system.cli.main", "init", str(tmp_path)],
             capture_output=True,
             text=True,
             cwd=Path(__file__).parent.parent.parent
         )
 
         result = subprocess.run(
-            [sys.executable, "-m", "research_system.cli.main", "v4-ingest",
+            [sys.executable, "-m", "research_system.cli.main", "ingest",
              "--workspace", str(tmp_path)],
             capture_output=True,
             text=True,
@@ -251,11 +251,11 @@ class TestV4NotImplemented:
         assert result.returncode == 0
         assert "no files found in inbox" in result.stdout.lower()
 
-    def test_v4_ingest_processes_file(self, tmp_path):
-        """Test v4-ingest processes a file from inbox."""
-        # First init a v4 workspace
+    def test_ingest_processes_file(self, tmp_path):
+        """Test ingest processes a file from inbox."""
+        # First init a workspace
         subprocess.run(
-            [sys.executable, "-m", "research_system.cli.main", "init", "--v4", str(tmp_path)],
+            [sys.executable, "-m", "research_system.cli.main", "init", str(tmp_path)],
             capture_output=True,
             text=True,
             cwd=Path(__file__).parent.parent.parent
@@ -267,7 +267,7 @@ class TestV4NotImplemented:
         test_file.write_text("Test strategy document")
 
         result = subprocess.run(
-            [sys.executable, "-m", "research_system.cli.main", "v4-ingest",
+            [sys.executable, "-m", "research_system.cli.main", "ingest",
              "--workspace", str(tmp_path), "--dry-run"],
             capture_output=True,
             text=True,
@@ -277,18 +277,18 @@ class TestV4NotImplemented:
         assert "processing" in result.stdout.lower()
         assert "test_strategy.txt" in result.stdout
 
-    def test_v4_verify_runs(self, tmp_path):
-        """Test v4-verify runs verification tests."""
-        # First init a v4 workspace
+    def test_verify_runs(self, tmp_path):
+        """Test verify runs verification tests."""
+        # First init a workspace
         subprocess.run(
-            [sys.executable, "-m", "research_system.cli.main", "init", "--v4", str(tmp_path)],
+            [sys.executable, "-m", "research_system.cli.main", "init", str(tmp_path)],
             capture_output=True,
             text=True,
             cwd=Path(__file__).parent.parent.parent
         )
 
         result = subprocess.run(
-            [sys.executable, "-m", "research_system.cli.main", "v4-verify",
+            [sys.executable, "-m", "research_system.cli.main", "verify",
              "--workspace", str(tmp_path), "STRAT-001"],
             capture_output=True,
             text=True,
@@ -297,18 +297,18 @@ class TestV4NotImplemented:
         # Returns 1 when strategy not found (which is expected)
         assert "Strategy ID required" in result.stdout or "not found" in result.stdout
 
-    def test_v4_validate_runs(self, tmp_path):
-        """Test v4-validate runs validation."""
-        # First init a v4 workspace
+    def test_validate_runs(self, tmp_path):
+        """Test validate runs validation."""
+        # First init a workspace
         subprocess.run(
-            [sys.executable, "-m", "research_system.cli.main", "init", "--v4", str(tmp_path)],
+            [sys.executable, "-m", "research_system.cli.main", "init", str(tmp_path)],
             capture_output=True,
             text=True,
             cwd=Path(__file__).parent.parent.parent
         )
 
         result = subprocess.run(
-            [sys.executable, "-m", "research_system.cli.main", "v4-validate",
+            [sys.executable, "-m", "research_system.cli.main", "validate",
              "--workspace", str(tmp_path), "STRAT-001"],
             capture_output=True,
             text=True,
@@ -317,18 +317,18 @@ class TestV4NotImplemented:
         # Returns error when strategy not found
         assert "Strategy ID required" in result.stdout or "not found" in result.stdout
 
-    def test_v4_learn_runs(self, tmp_path):
-        """Test v4-learn extracts learnings."""
-        # First init a v4 workspace
+    def test_learn_runs(self, tmp_path):
+        """Test learn extracts learnings."""
+        # First init a workspace
         subprocess.run(
-            [sys.executable, "-m", "research_system.cli.main", "init", "--v4", str(tmp_path)],
+            [sys.executable, "-m", "research_system.cli.main", "init", str(tmp_path)],
             capture_output=True,
             text=True,
             cwd=Path(__file__).parent.parent.parent
         )
 
         result = subprocess.run(
-            [sys.executable, "-m", "research_system.cli.main", "v4-learn",
+            [sys.executable, "-m", "research_system.cli.main", "learn",
              "--workspace", str(tmp_path), "STRAT-001"],
             capture_output=True,
             text=True,
@@ -337,39 +337,39 @@ class TestV4NotImplemented:
         # Returns error when strategy not found
         assert "Strategy ID required" in result.stdout or "not found" in result.stdout
 
-    def test_v4_status_not_implemented(self, tmp_path):
-        """Test v4-status prints not implemented message."""
-        # First init a v4 workspace
+    def test_status_works(self, tmp_path):
+        """Test status shows workspace status."""
+        # First init a workspace
         subprocess.run(
-            [sys.executable, "-m", "research_system.cli.main", "init", "--v4", str(tmp_path)],
+            [sys.executable, "-m", "research_system.cli.main", "init", str(tmp_path)],
             capture_output=True,
             text=True,
             cwd=Path(__file__).parent.parent.parent
         )
 
         result = subprocess.run(
-            [sys.executable, "-m", "research_system.cli.main", "v4-status",
+            [sys.executable, "-m", "research_system.cli.main", "status",
              "--workspace", str(tmp_path)],
             capture_output=True,
             text=True,
             cwd=Path(__file__).parent.parent.parent
         )
         assert result.returncode == 0
-        assert "Research-Kit V4" in result.stdout
+        assert "Research-Kit" in result.stdout
         assert "Workspace" in result.stdout
 
-    def test_v4_list_works(self, tmp_path):
-        """Test v4-list works on empty workspace."""
-        # First init a v4 workspace
+    def test_list_works(self, tmp_path):
+        """Test list works on empty workspace."""
+        # First init a workspace
         subprocess.run(
-            [sys.executable, "-m", "research_system.cli.main", "init", "--v4", str(tmp_path)],
+            [sys.executable, "-m", "research_system.cli.main", "init", str(tmp_path)],
             capture_output=True,
             text=True,
             cwd=Path(__file__).parent.parent.parent
         )
 
         result = subprocess.run(
-            [sys.executable, "-m", "research_system.cli.main", "v4-list",
+            [sys.executable, "-m", "research_system.cli.main", "list",
              "--workspace", str(tmp_path)],
             capture_output=True,
             text=True,
@@ -378,18 +378,18 @@ class TestV4NotImplemented:
         assert result.returncode == 0
         assert "No strategies" in result.stdout
 
-    def test_v4_show_not_found(self, tmp_path):
-        """Test v4-show returns error for non-existent strategy."""
-        # First init a v4 workspace
+    def test_show_not_found(self, tmp_path):
+        """Test show returns error for non-existent strategy."""
+        # First init a workspace
         subprocess.run(
-            [sys.executable, "-m", "research_system.cli.main", "init", "--v4", str(tmp_path)],
+            [sys.executable, "-m", "research_system.cli.main", "init", str(tmp_path)],
             capture_output=True,
             text=True,
             cwd=Path(__file__).parent.parent.parent
         )
 
         result = subprocess.run(
-            [sys.executable, "-m", "research_system.cli.main", "v4-show",
+            [sys.executable, "-m", "research_system.cli.main", "show",
              "--workspace", str(tmp_path), "STRAT-001"],
             capture_output=True,
             text=True,
@@ -400,24 +400,24 @@ class TestV4NotImplemented:
 
 
 # =============================================================================
-# TEST INIT --V4
+# TEST INIT
 # =============================================================================
 
 
-class TestV4Init:
-    """Test research init --v4 command."""
+class TestInit:
+    """Test research init command."""
 
-    def test_init_v4_creates_workspace(self, tmp_path):
-        """Test init --v4 creates V4 workspace structure."""
+    def test_init_creates_workspace(self, tmp_path):
+        """Test init creates workspace structure."""
         result = subprocess.run(
-            [sys.executable, "-m", "research_system.cli.main", "init", "--v4", str(tmp_path)],
+            [sys.executable, "-m", "research_system.cli.main", "init", str(tmp_path)],
             capture_output=True,
             text=True,
             cwd=Path(__file__).parent.parent.parent
         )
 
         assert result.returncode == 0
-        assert "Initialized V4 workspace" in result.stdout
+        assert "Initialized workspace" in result.stdout
         assert (tmp_path / "research-kit.yaml").exists()
         assert (tmp_path / "inbox").exists()
         assert (tmp_path / "strategies" / "pending").exists()
@@ -428,10 +428,10 @@ class TestV4Init:
         assert (tmp_path / "learnings").exists()
         assert (tmp_path / "ideas").exists()
 
-    def test_init_v4_shows_next_steps(self, tmp_path):
-        """Test init --v4 shows next steps."""
+    def test_init_shows_next_steps(self, tmp_path):
+        """Test init shows next steps."""
         result = subprocess.run(
-            [sys.executable, "-m", "research_system.cli.main", "init", "--v4", str(tmp_path)],
+            [sys.executable, "-m", "research_system.cli.main", "init", str(tmp_path)],
             capture_output=True,
             text=True,
             cwd=Path(__file__).parent.parent.parent
@@ -439,15 +439,15 @@ class TestV4Init:
 
         assert result.returncode == 0
         assert "Next steps" in result.stdout
-        assert "v4-ingest" in result.stdout
-        assert "v4-verify" in result.stdout
-        assert "v4-validate" in result.stdout
+        assert "ingest" in result.stdout
+        assert "verify" in result.stdout
+        assert "validate" in result.stdout
 
-    def test_init_v4_already_exists(self, tmp_path):
-        """Test init --v4 fails if workspace already exists."""
+    def test_init_already_exists(self, tmp_path):
+        """Test init fails if workspace already exists."""
         # First init
         subprocess.run(
-            [sys.executable, "-m", "research_system.cli.main", "init", "--v4", str(tmp_path)],
+            [sys.executable, "-m", "research_system.cli.main", "init", str(tmp_path)],
             capture_output=True,
             text=True,
             cwd=Path(__file__).parent.parent.parent
@@ -455,7 +455,7 @@ class TestV4Init:
 
         # Second init should fail
         result = subprocess.run(
-            [sys.executable, "-m", "research_system.cli.main", "init", "--v4", str(tmp_path)],
+            [sys.executable, "-m", "research_system.cli.main", "init", str(tmp_path)],
             capture_output=True,
             text=True,
             cwd=Path(__file__).parent.parent.parent
@@ -464,11 +464,11 @@ class TestV4Init:
         assert result.returncode == 1
         assert "already exists" in result.stdout
 
-    def test_init_v4_force_reinitializes(self, tmp_path):
-        """Test init --v4 --force reinitializes workspace."""
+    def test_init_force_reinitializes(self, tmp_path):
+        """Test init --force reinitializes workspace."""
         # First init
         subprocess.run(
-            [sys.executable, "-m", "research_system.cli.main", "init", "--v4", str(tmp_path)],
+            [sys.executable, "-m", "research_system.cli.main", "init", str(tmp_path)],
             capture_output=True,
             text=True,
             cwd=Path(__file__).parent.parent.parent
@@ -476,14 +476,14 @@ class TestV4Init:
 
         # Force reinit should succeed
         result = subprocess.run(
-            [sys.executable, "-m", "research_system.cli.main", "init", "--v4", "--force", str(tmp_path)],
+            [sys.executable, "-m", "research_system.cli.main", "init", "--force", str(tmp_path)],
             capture_output=True,
             text=True,
             cwd=Path(__file__).parent.parent.parent
         )
 
         assert result.returncode == 0
-        assert "Initialized V4 workspace" in result.stdout
+        assert "Initialized workspace" in result.stdout
 
 
 # =============================================================================
@@ -491,21 +491,21 @@ class TestV4Init:
 # =============================================================================
 
 
-class TestV4WorkspaceFlag:
-    """Test that --workspace flag works for V4 commands."""
+class TestWorkspaceFlag:
+    """Test that --workspace flag works for commands."""
 
-    def test_v4_ingest_accepts_workspace_flag(self, tmp_path):
-        """Test v4-ingest accepts --workspace flag."""
-        # First init a v4 workspace
+    def test_ingest_accepts_workspace_flag(self, tmp_path):
+        """Test ingest accepts --workspace flag."""
+        # First init a workspace
         subprocess.run(
-            [sys.executable, "-m", "research_system.cli.main", "init", "--v4", str(tmp_path)],
+            [sys.executable, "-m", "research_system.cli.main", "init", str(tmp_path)],
             capture_output=True,
             text=True,
             cwd=Path(__file__).parent.parent.parent
         )
 
         result = subprocess.run(
-            [sys.executable, "-m", "research_system.cli.main", "v4-ingest",
+            [sys.executable, "-m", "research_system.cli.main", "ingest",
              "--workspace", str(tmp_path)],
             capture_output=True,
             text=True,
@@ -515,18 +515,18 @@ class TestV4WorkspaceFlag:
         assert result.returncode == 0
         assert str(tmp_path) in result.stdout
 
-    def test_v4_ingest_accepts_short_workspace_flag(self, tmp_path):
-        """Test v4-ingest accepts -w flag."""
-        # First init a v4 workspace
+    def test_ingest_accepts_short_workspace_flag(self, tmp_path):
+        """Test ingest accepts -w flag."""
+        # First init a workspace
         subprocess.run(
-            [sys.executable, "-m", "research_system.cli.main", "init", "--v4", str(tmp_path)],
+            [sys.executable, "-m", "research_system.cli.main", "init", str(tmp_path)],
             capture_output=True,
             text=True,
             cwd=Path(__file__).parent.parent.parent
         )
 
         result = subprocess.run(
-            [sys.executable, "-m", "research_system.cli.main", "v4-ingest",
+            [sys.executable, "-m", "research_system.cli.main", "ingest",
              "-w", str(tmp_path)],
             capture_output=True,
             text=True,
@@ -536,18 +536,18 @@ class TestV4WorkspaceFlag:
         assert result.returncode == 0
         assert str(tmp_path) in result.stdout
 
-    def test_v4_status_accepts_workspace_flag(self, tmp_path):
-        """Test v4-status accepts --workspace flag."""
-        # First init a v4 workspace
+    def test_status_accepts_workspace_flag(self, tmp_path):
+        """Test status accepts --workspace flag."""
+        # First init a workspace
         subprocess.run(
-            [sys.executable, "-m", "research_system.cli.main", "init", "--v4", str(tmp_path)],
+            [sys.executable, "-m", "research_system.cli.main", "init", str(tmp_path)],
             capture_output=True,
             text=True,
             cwd=Path(__file__).parent.parent.parent
         )
 
         result = subprocess.run(
-            [sys.executable, "-m", "research_system.cli.main", "v4-status",
+            [sys.executable, "-m", "research_system.cli.main", "status",
              "--workspace", str(tmp_path)],
             capture_output=True,
             text=True,
@@ -559,25 +559,25 @@ class TestV4WorkspaceFlag:
 
 
 # =============================================================================
-# TEST V4 CONFIG COMMAND
+# TEST CONFIG COMMAND
 # =============================================================================
 
 
-class TestV4Config:
-    """Test v4-config command."""
+class TestConfigCommand:
+    """Test config command."""
 
-    def test_v4_config_shows_configuration(self, tmp_path):
-        """Test v4-config shows configuration values."""
-        # First init a v4 workspace
+    def test_config_shows_configuration(self, tmp_path):
+        """Test config shows configuration values."""
+        # First init a workspace
         subprocess.run(
-            [sys.executable, "-m", "research_system.cli.main", "init", "--v4", str(tmp_path)],
+            [sys.executable, "-m", "research_system.cli.main", "init", str(tmp_path)],
             capture_output=True,
             text=True,
             cwd=Path(__file__).parent.parent.parent
         )
 
         result = subprocess.run(
-            [sys.executable, "-m", "research_system.cli.main", "v4-config",
+            [sys.executable, "-m", "research_system.cli.main", "config",
              "--workspace", str(tmp_path)],
             capture_output=True,
             text=True,
@@ -590,18 +590,18 @@ class TestV4Config:
         assert "max_drawdown" in result.stdout
         assert "research-kit.yaml" in result.stdout
 
-    def test_v4_config_validate_passes(self, tmp_path):
-        """Test v4-config --validate passes for valid config."""
-        # First init a v4 workspace
+    def test_config_validate_passes(self, tmp_path):
+        """Test config --validate passes for valid config."""
+        # First init a workspace
         subprocess.run(
-            [sys.executable, "-m", "research_system.cli.main", "init", "--v4", str(tmp_path)],
+            [sys.executable, "-m", "research_system.cli.main", "init", str(tmp_path)],
             capture_output=True,
             text=True,
             cwd=Path(__file__).parent.parent.parent
         )
 
         result = subprocess.run(
-            [sys.executable, "-m", "research_system.cli.main", "v4-config",
+            [sys.executable, "-m", "research_system.cli.main", "config",
              "--workspace", str(tmp_path), "--validate"],
             capture_output=True,
             text=True,
@@ -617,13 +617,13 @@ class TestV4Config:
 # =============================================================================
 
 
-class TestV4ErrorHandling:
-    """Test error handling for V4 commands."""
+class TestErrorHandling:
+    """Test error handling for commands."""
 
-    def test_v4_ingest_requires_initialized_workspace(self, tmp_path):
-        """Test v4-ingest fails if workspace not initialized."""
+    def test_ingest_requires_initialized_workspace(self, tmp_path):
+        """Test ingest fails if workspace not initialized."""
         result = subprocess.run(
-            [sys.executable, "-m", "research_system.cli.main", "v4-ingest",
+            [sys.executable, "-m", "research_system.cli.main", "ingest",
              "--workspace", str(tmp_path)],
             capture_output=True,
             text=True,
@@ -631,12 +631,12 @@ class TestV4ErrorHandling:
         )
 
         assert result.returncode == 1
-        assert "not initialized" in result.stdout.lower() or "init --v4" in result.stdout
+        assert "not initialized" in result.stdout.lower() or "init" in result.stdout
 
-    def test_v4_status_requires_initialized_workspace(self, tmp_path):
-        """Test v4-status fails if workspace not initialized."""
+    def test_status_requires_initialized_workspace(self, tmp_path):
+        """Test status fails if workspace not initialized."""
         result = subprocess.run(
-            [sys.executable, "-m", "research_system.cli.main", "v4-status",
+            [sys.executable, "-m", "research_system.cli.main", "status",
              "--workspace", str(tmp_path)],
             capture_output=True,
             text=True,
@@ -644,12 +644,12 @@ class TestV4ErrorHandling:
         )
 
         assert result.returncode == 1
-        assert "not initialized" in result.stdout.lower() or "init --v4" in result.stdout
+        assert "not initialized" in result.stdout.lower() or "init" in result.stdout
 
-    def test_v4_config_requires_initialized_workspace(self, tmp_path):
-        """Test v4-config fails if workspace not initialized."""
+    def test_config_requires_initialized_workspace(self, tmp_path):
+        """Test config fails if workspace not initialized."""
         result = subprocess.run(
-            [sys.executable, "-m", "research_system.cli.main", "v4-config",
+            [sys.executable, "-m", "research_system.cli.main", "config",
              "--workspace", str(tmp_path)],
             capture_output=True,
             text=True,
@@ -657,4 +657,4 @@ class TestV4ErrorHandling:
         )
 
         assert result.returncode == 1
-        assert "not initialized" in result.stdout.lower() or "init --v4" in result.stdout
+        assert "not initialized" in result.stdout.lower() or "init" in result.stdout
